elixfix v0.0.1-dev FSession.Support

Support pure functions to work with Sessions

Functions

confirm_login_msg(heart_beat)

reject_msg(description, original_message)

It will return a set of tags for session message reject

You have to provide:

Description

Original message

Reject code

5 : Incorrect value

6 : Incorrect data format for a tag value

9 : CompID problem

example:

iex> original_message = %{:MsgType => "3", :MsgSeqNum => "101", :RefMsgType => "B"}
iex> FSession.Support.reject_msg("Testing reject", original_message)
%{
    :MsgType => "3",
    :RefMsgType => "B",
    :RefSeqNum => "101",
    :Text => "Testing reject"
}

request_resend(begin_seq_num, end_seq_num)

It will return a set of tags for RequestResend

You have to provide:

BeginSeqNum

EndSeqNum

example:

iex> FSession.Support.request_resend("101", "314")
%{:MsgType => "2",
  :BeginSeqNo => "101",
  :EndSeqNo => "314"}