elixfix v0.0.1-dev FTags

Functions to work and convert tags atom, int, string

atom -> int

atom -> string

int -> atom

Functions

get_atom(unknown)

Convert from int to atom

iex> FTags.get_atom(8)
:BeginString

If tag is not known, it will return the received integer

get_name(unknown)

Convert from atom to string

iex> FTags.get_name(:BeginString)
"BeginString(8)"

If atom is not know, it will return atom->string

get_num(arg1)

Convert from atom to int

iex> FTags.get_num(:BeginString)
8

If atom is not known, it will return 0